Cleanup, fixed scraper_action test visibility
